requirements.txt
websockets
aiohttp
discord.py>=1.7.2
